requirements.txt
discord==1.0.1
pandas==1.3.3
pendulum==2.1.2
pytz==2021.1
python-decouple==3.5
xlrd >= 1.0.0